requirements.txt
click~=8.1.7
pyserial~=3.5